server/pyproject.toml
[project]
name = "tv-commercial-detector-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.0",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "opencv-python-headless>=4.13.0.92",
    "imagehash>=4.3.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"tv_commercial_detector" = ["templates/*", "prompt/*", "prompt/logos/*"]

[tool.ruff]
src = ["src"]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]

[tool.ruff.format]
# defaults are fine

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
markers = ["integration: marks tests as integration tests (deselect with '-m not integration')"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.6",
]
